+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     * New: Add `initialize_setting()`, `allowed_options()`, `display_option()`, and `plugin_action_links()`
 * Change: Use default time format if parameter or filter attempts to configure an empty string or non-string value
 * Fix: Add support for the 'T' timezone format character to the dynamic clock (support for which was removed from Moment.js awhile ago)
+* Change: Switch to use of `wp_add_inline_script()` instead of `wp_localize_script()`
 * Change: Note compatibility through WP 5.7+
 * Change: Update URLs to PHP documentation for datetime formatting
 * Change: Update copyright date (2021)